Make parser less strict.
This commit introduces a new parser which is less particular about the form of terms which it is asked to parse. It is backwards compatible with the old parser, but makes it possible to write expressions which are less over-parenthesized through the introduction of two new rules. The first is that the body of an abstraction extends as far to the right as possible. The second is that application is taken to be left-associative unless otherwise specified.